Adding the BrandingByReferer middleware in the right place. Fixing a couple of linting errors
