+++ lms/djangoapps/branding_by_referer/__init__.py
@@ -1,2 +1,1 @@
-
 from .branding_by_referer import *
